+++ egi_timer.h
@@ -39,7 +39,7 @@ void tm_start_egitick(void);
 long long unsigned int tm_get_tickcount(void);
 void tm_delayms(unsigned long ms);/* !!! To be abandoned !!!! Not good! */
 bool tm_pulseus(long long unsigned int gap, unsigned int t); /* gap(us) */
-unsigned int tm_diffus(struct timeval t_start, struct timeval t_end);
+unsigned long tm_diffus(struct timeval t_start, struct timeval t_end);
 int tm_signed_diffms(struct timeval tm_start, struct timeval tm_end);
 void egi_sleep(unsigned char fd, unsigned int s, unsigned int ms);
